+++ NEWS.md
@@ -10,6 +10,7 @@ plumber 1.0.0.9999 Development version
 * Guess OpenApi response content type from serializer (@meztez #684)
 
 * Passing `edit = TRUE` to `plumb_api()` will open the API source file (#699)
+* Allow for spaces in `@apiTag` and `@tag` when tag is surrended by single or double quotes (#685)
 
 * OpenAPI Specification can be set using a file path. (@meztez #696)